Archive/ArchiveVer2_20251119/requirements.txt
fastapi
uvicorn[standard]
requests
SQLAlchemy
python-jose[cryptography]
